+++ clownaudio/decoding/low-level-decoder_selector.c
@@ -92,7 +92,7 @@ LowLevelDecoderSelector* LowLevelDecoderSelector_Create(const unsigned char *dat
 {
 	if (data != NULL)
 	{
-		for (unsigned int i = 0; i < sizeof(decoder_functions) / sizeof(decoder_functions[0]); ++i)
+		for (size_t i = 0; i < sizeof(decoder_functions) / sizeof(decoder_functions[0]); ++i)
 		{
 			void *decoder = decoder_functions[i].Create(data, data_size, info);
